pyproject.toml
[tool.poetry]
name = "expensesharing"
version = "0.1.0"
description = "This app is used to share your expenses in your firiend"
authors = ["avijeetjha77 <avijeetjha@rirabh.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.1"
django-rest-passwordreset = "1.3.0"
django-redis = "^5.4.0"
django-split-settings = "^1.3.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.7.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
